task execution log in the worker
