+++ client/src/test/java/com/netflix/conductor/client/worker/TestPropertyFactory.java
@@ -5,10 +5,14 @@ package com.netflix.conductor.client.worker;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
 import org.junit.BeforeClass;
 import org.junit.Test;
+
+import com.netflix.conductor.common.metadata.tasks.Task;
+import com.netflix.conductor.common.metadata.tasks.TaskResult;
 
 /**
  * @author Viren
@@ -26,12 +30,25 @@ public class TestPropertyFactory {
 		System.setProperty("conductor.worker.workerA.paused", "true");
 		
 		System.setProperty("conductor.worker.workerB.batchSize", "84");
+		
+		System.setProperty("conductor.worker.Test.paused", "true");
+	}
+	
+	@Test
+	public void testIdentity(){
+		Worker worker = Worker.create("Test2", (Task task)->{
+			return new TaskResult(task);
+		});
+		assertNotNull(worker.getIdentity());
+		boolean paused = worker.paused();
+		assertFalse("Paused? " + paused, paused);
 	}
 	
 	@Test
 	public void test() {
 		
-		assertEquals(2, PropertyFactory.getInteger("workerB", "pollingInterval", 100).intValue());
+		int val = PropertyFactory.getInteger("workerB", "pollingInterval", 100).intValue();
+		assertEquals("got: " + val, 2, val);
 		assertEquals(100, PropertyFactory.getInteger("workerB", "propWithoutValue", 100).intValue());
 		
 		assertFalse(PropertyFactory.getBoolean("workerB", "paused", true));		//Global value set to 'false'
@@ -42,6 +59,15 @@ public class TestPropertyFactory {
 		assertEquals(84, PropertyFactory.getInteger("workerB", "batchSize", 42).intValue());	//WorkerB's value set to 84
 		
 
+	}
+	
+	@Test
+	public void testProperty() {
+		Worker worker = Worker.create("Test", (Task task)->{
+			return new TaskResult(task);
+		});
+		boolean paused = worker.paused();
+		assertTrue("Paused? " + paused, paused);
 	}
 	
 }
